improved slurp.rb

+++ bin/slurp.rb
@@ -10,16 +10,19 @@ dump_path = ARGV[0]
 db_name   = ARGV[1] || File.basename(Dir.pwd)
 
 # scp, and uncompress the database dump
-fail "Failed to copy." unless system("scp mostfit.in:#{dump_path} .")
-puts "Uncompressing ..."
-file_name = dump_path.split("/")[-1]
-fail "Failed to uncompress" unless system("bunzip2 #{file_name}")
-dump_path = File.basename(dump_path, '.bz2')
-
+sql_filename = File.basename(dump_path,".bz2")
+bz2_filename = dump_path.split("/")[-1]
+unless (File.exists?(sql_filename) or File.exists?(bz2_filename))
+  fail "Failed to copy." unless system("scp mostfit.in:#{dump_path} .")
+end
+unless File.exists?(sql_filename)
+  puts "Uncompressinging #{bz2_filename} ..."
+  fail "Failed to uncompress" unless system("bunzip2 #{bz2_filename}")
+end
 # backup - defaults to no
 print "Backup the current database? [y/N]: "
 choice = $stdin.gets.chomp.downcase
 system('./bin/dump.rb') if choice == 'y'
 
 # now load the db
-fail "Failed to load database dump." unless system("mysql -p -u root #{db_name} < #{dump_path}")
+fail "Failed to load database dump." unless system("mysql -p -u root #{db_name} < #{sql_filename}")
